+++ src/renderer/viz/expressions/classification/GlobalEqIntervals.js
@@ -2,6 +2,7 @@ import Classifier from './Classifier';
 import { checkExactNumberOfArguments } from '../utils';
 import { CLUSTER_FEATURE_COUNT } from '../../../../constants/metadata';
 import CartoValidationError, { CartoValidationErrorTypes } from '../../../../errors/carto-validation-error';
+import { number } from '../../expressions';
 
 /**
  * Classify `input` by using the equal intervals method with `n` buckets.
@@ -36,7 +37,6 @@ export default class GlobalEqIntervals extends Classifier {
 
     _bindMetadata (metadata) {
         super._bindMetadata(metadata);
-
         this._updateBreakpointsWith(metadata);
     }
 
@@ -50,8 +50,8 @@ export default class GlobalEqIntervals extends Classifier {
 
         const name = this.input.name;
         const { min, max } = metadata.stats(name);
-        this.min = min;
-        this.max = max;
+        this.min = number(min);
+        this.max = number(max);
 
         this.breakpoints.map((breakpoint, index) => {
             const p = (index + 1) / this.numCategories;